TD 2/app_faker/requirements.txt
mysql-connector-python==8.0.33
Faker==19.6.2
